Replace PowerMock from AlluxioFileInStreamTest
